+++ Molulip.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27130.2024
+VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Molulip", "Molulip\Molulip.csproj", "{29895803-11AB-4D1F-8C10-62E1951025A6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Molulip", "Molulip\Molulip.csproj", "{29895803-11AB-4D1F-8C10-62E1951025A6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Molulip.Import", "Molulip.Import\Molulip.Import.csproj", "{6672EFD6-7C11-4DAC-9945-07998EAAC942}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{29895803-11AB-4D1F-8C10-62E1951025A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{29895803-11AB-4D1F-8C10-62E1951025A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{29895803-11AB-4D1F-8C10-62E1951025A6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6672EFD6-7C11-4DAC-9945-07998EAAC942}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6672EFD6-7C11-4DAC-9945-07998EAAC942}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6672EFD6-7C11-4DAC-9945-07998EAAC942}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6672EFD6-7C11-4DAC-9945-07998EAAC942}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
