+++ Molulip/Controllers/HomeController.cs
@@ -1,14 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using Molulip.Models;
-using System.IO;
-using OfficeOpenXml;
-using System.Text;
-using Molulip.Extentions;
 
 namespace Molulip.Controllers
 {
@@ -16,8 +8,6 @@ namespace Molulip.Controllers
     {
         public IActionResult Index()
         {
-            var x = GetMeals();
-
             return View();
         }
 
@@ -38,35 +28,6 @@ namespace Molulip.Controllers
         public IActionResult Error()
         {
             return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-        }
-
-        public List<Meal> GetMeals()
-        {
-            var mealList = new List<Meal>();
-
-            try
-            {
-                var baseDirectory = System.AppDomain.CurrentDomain.BaseDirectory;
-                var fileName = "Feb.xlsx";
-
-                var filePath = Path.GetFullPath(Path.Combine(baseDirectory, fileName));
-
-                FileInfo file = new FileInfo(filePath);
-
-                using (ExcelPackage package = new ExcelPackage(file))
-                {
-                    ExcelWorksheet worksheet = package.Workbook.Worksheets[1];
-
-                    mealList = worksheet.ConvertSheetToObjects<Meal>().ToList();
-                }
-            }
-            catch (Exception ex)
-            {
-
-                throw ex;
-            }
-
-            return mealList;
         }
     }
 }
